+++ desktop/php/jollymec.php
@@ -49,6 +49,13 @@ $eqLogics = eqLogic::byType($plugin->getId());
         <ul class="nav nav-tabs" role="tablist">
             <li role="presentation"><a class="eqLogicAction cursor" aria-controls="home" role="tab" data-action="returnToThumbnailDisplay"><i class="fas fa-arrow-circle-left"></i></a></li>
             <li role="presentation" class="active"><a href="#eqlogictab" aria-controls="home" role="tab" data-toggle="tab"><i class="fa fa-tachometer"></i> {{Poêle}}</a></li>
+            <?php
+            if (config::byKey('handle_chronothermostat', 'jollymec')) {
+            ?>
+            <li role="presentation"><a href="#chronotab" aria-controls="profile" role="tab" data-toggle="tab"><i class="fa fa-clock"></i> {{ChronoThermostat}}</a></li>
+            <?php
+            }
+            ?>
             <li role="presentation"><a href="#commandtab" aria-controls="profile" role="tab" data-toggle="tab"><i class="fa fa-list-alt"></i> {{Commandes}}</a></li>
         </ul>
         <div class="tab-content" style="height:calc(100% - 50px);overflow:auto;overflow-x: hidden;">
@@ -66,12 +73,12 @@ $eqLogics = eqLogic::byType($plugin->getId());
                         <div class="form-group">
                             <label class="col-sm-3 control-label" >{{Objet parent}}</label>
                             <div class="col-sm-3">
-                                <select id="sel_object" class="eqLogicAttr form-control" data-l1key="object_id">
+                                <select id="sel_object" class=" form-control" data-l1key="object_id">
                                     <option value="">{{Aucun}}</option>
                                     <?php
-                                        foreach (jeeObject::all() as $object) {
-                                            echo '<option value="' . $object->getId() . '">' . $object->getName() . '</option>';
-                                        }
+                                    foreach (jeeObject::all() as $object) {
+                                        echo '<option value="' . $object->getId() . '">' . $object->getName() . '</option>';
+                                    }
                                     ?>
                                 </select>
                             </div>
@@ -107,6 +114,296 @@ $eqLogics = eqLogic::byType($plugin->getId());
                     </fieldset>
                 </form>
             </div>
+            <?php
+            if (config::byKey('handle_chronothermostat', 'jollymec')) {
+            ?>
+            <div role="tabpanel" class="tab-pane" id="chronotab">
+                <br />
+                <form class="form-horizontal">
+                    <fieldset>
+                        <div class="form-group">
+                            <label class="col-sm-3 control-label">{{Activer}}</label>
+                            <div class="col-sm-9">
+                                <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[enableChrono]" />{{Activer}}</label>
+                                <input type="hidden" name="chrono[programs]" value="4" />
+                            </div>
+                        </div>
+                    </fieldset>
+                    <div id="chronoprograms" class="panel-group">
+                        <div class="mode panel panel-default">
+                            <div class="panel-heading"><h3 class="panel-title"><a href="#collapsechonoprogram1" class="accordion-toggle" data-toggle="collapse" data-parent="#chronoprograms" aria-expanded="true">{{Programme 1}}</a></h3></div>
+                            <div id="collapsechonoprogram1" class="panel-collapse collapse in" aria-expanded="true" style="">
+                                <div class="panel-body">
+                                    <div class="well">
+                                        <fieldset>
+                                            <div class="form-group">
+                                                <div class="col-sm-9">
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][0]" />{{Lundi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][1]" />{{Mardi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][2]" />{{Mercredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][3]" />{{Jeudi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][4]" />{{Vendredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][5]" />{{Samedi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[1][gg][6]" />{{Dimanche}}</label>
+                                                </div>
+                                            </div>
+
+                                            <div class="form-group">
+                                                <label class="col-sm-1 control-label">{{Allumage}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[1][start]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Arrêt}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[1][stop]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Puissance}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[1][setPower]">
+                                                        <?php
+                                                            for ($i = 0; $i < 6; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Thermostat}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[1][setTemp]">
+                                                        <?php
+                                                            for ($i = 7; $i < 41; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '°</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                            </div>
+                                        </fieldset>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="mode panel panel-default">
+                            <div class="panel-heading"><h3 class="panel-title"><a href="#collapsechonoprogram2" class="accordion-toggle" data-toggle="collapse" data-parent="#chronoprograms" aria-expanded="true">{{Programme 2}}</a></h3></div>
+                            <div id="collapsechonoprogram2" class="panel-collapse collapse in" aria-expanded="true" style="">
+                                <div class="panel-body">
+                                    <div class="well">
+                                        <fieldset>
+                                            <div class="form-group">
+                                                <div class="col-sm-9">
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][0]" />{{Lundi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][1]" />{{Mardi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][2]" />{{Mercredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][3]" />{{Jeudi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][4]" />{{Vendredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][5]" />{{Samedi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[2][gg][6]" />{{Dimanche}}</label>
+                                                </div>
+                                            </div>
+
+                                            <div class="form-group">
+                                                <label class="col-sm-1 control-label">{{Allumage}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[2][start]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Arrêt}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[2][stop]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Puissance}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[2][setPower]">
+                                                        <?php
+                                                            for ($i = 0; $i < 6; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Thermostat}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[2][setTemp]">
+                                                        <?php
+                                                            for ($i = 7; $i < 41; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '°</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                            </div>
+                                        </fieldset>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="mode panel panel-default">
+                            <div class="panel-heading"><h3 class="panel-title"><a href="#collapsechonoprogram3" class="accordion-toggle" data-toggle="collapse" data-parent="#chronoprograms" aria-expanded="true">{{Programme 3}}</a></h3></div>
+                            <div id="collapsechonoprogram3" class="panel-collapse collapse in" aria-expanded="true" style="">
+                                <div class="panel-body">
+                                    <div class="well">
+                                        <fieldset>
+                                            <div class="form-group">
+                                                <div class="col-sm-9">
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][0]" />{{Lundi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][1]" />{{Mardi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][2]" />{{Mercredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][3]" />{{Jeudi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][4]" />{{Vendredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][5]" />{{Samedi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[3][gg][6]" />{{Dimanche}}</label>
+                                                </div>
+                                            </div>
+
+                                            <div class="form-group">
+                                                <label class="col-sm-1 control-label">{{Allumage}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[3][start]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Arrêt}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[3][stop]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Puissance}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[3][setPower]">
+                                                        <?php
+                                                            for ($i = 0; $i < 6; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Thermostat}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[3][setTemp]">
+                                                        <?php
+                                                            for ($i = 7; $i < 41; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '°</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                            </div>
+                                        </fieldset>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="mode panel panel-default">
+                            <div class="panel-heading"><h3 class="panel-title"><a href="#collapsechonoprogram4" class="accordion-toggle" data-toggle="collapse" data-parent="#chronoprograms" aria-expanded="true">{{Programme 4}}</a></h3></div>
+                            <div id="collapsechonoprogram4" class="panel-collapse collapse in" aria-expanded="true" style="">
+                                <div class="panel-body">
+                                    <div class="well">
+                                        <fieldset>
+                                            <div class="form-group">
+                                                <div class="col-sm-9">
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][0]" />{{Lundi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][1]" />{{Mardi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][2]" />{{Mercredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][3]" />{{Jeudi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][4]" />{{Vendredi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][5]" />{{Samedi}}</label>
+                                                    <label class="checkbox-inline"><input type="checkbox" class="" name="chrono[4][gg][6]" />{{Dimanche}}</label>
+                                                </div>
+                                            </div>
+
+                                            <div class="form-group">
+                                                <label class="col-sm-1 control-label">{{Allumage}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[4][start]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Arrêt}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[4][stop]">
+                                                        <option value="144">{{OFF}}</option>
+                                                        <?php
+                                                            for ($i = 0; $i < 144; $i++) {
+                                                                echo '<option value="' . $i . '">' . date('H:i', mktime('0', '0', $i*10*60, '1', '1', '1970')) . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Puissance}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[4][setPower]">
+                                                        <?php
+                                                            for ($i = 0; $i < 6; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                                <label class="col-sm-1 control-label">{{Thermostat}}</label>
+                                                <div class="col-sm-1">
+                                                    <select class=" form-control input-sm" name="chrono[4][setTemp]">
+                                                        <?php
+                                                            for ($i = 7; $i < 41; $i++) {
+                                                                echo '<option value="' . $i . '">' . $i . '°</option>';
+                                                            }
+                                                        ?>
+                                                    </select>
+                                                </div>
+                                            </div>
+                                        </fieldset>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </form>
+            </div>
+            <?php
+            }
+            ?>
             <div role="tabpanel" class="tab-pane" id="commandtab">
                 <table id="table_cmd" class="table table-bordered table-condensed">
                     <thead>